Add partner key selection to agency dashboard
