+++ client/jetpack-cloud/sections/partner-portal/utils.ts
@@ -42,7 +42,6 @@ export function getLicenseState(
  * This is a hack around TypeScript's poor support of enums as types.
  *
  * @example const enumMember = valueToEnum< SomeEnumType >( SomeEnumType, 'foo', SomeEnumType.SomeMember );
- *
  * @template T
  * @param {Record< string, * >} enumType Enum type to search in.
  * @param {*} value The enum value we are looking to get the member for.
